requirements.txt
Pillow
pyqt5
pystray
winotify
schedule